+++ libs/cardano-protocol-tpraos/testlib/Test/Cardano/Protocol/Binary/RoundTrip.hs
@@ -6,13 +6,12 @@
 
 module Test.Cardano.Protocol.Binary.RoundTrip (roundTripBlockSpec) where
 
-import Cardano.Ledger.Binary (Annotator, DecCBOR, EncCBOR, EncCBORGroup)
+import Cardano.Ledger.Binary (Annotator, DecCBOR, EncCBOR, ToCBOR)
 import Cardano.Ledger.Block (Block)
 import Cardano.Ledger.Core
 import Data.Typeable
 import qualified Test.Cardano.Base.QuickCheck as BaseQC
 import Test.Cardano.Ledger.Common
-import Test.Cardano.Ledger.Core.Binary.Annotator ()
 import Test.Cardano.Ledger.Core.Binary.RoundTrip
 import Test.Cardano.Protocol.Binary.Annotator ()
 import Test.Cardano.Protocol.TPraos.Arbitrary ()
@@ -21,18 +20,17 @@ roundTripBlockSpec ::
   forall h era.
   ( Eq h
   , Show h
-  , DecCBOR h
-  , DecCBOR (Annotator h)
-  , EncCBOR h
   , EraBlockBody era
   , Arbitrary (Block h era)
-  , DecCBOR (BlockBody era)
-  , EncCBORGroup (BlockBody era)
+  , EncCBOR (Block h era)
+  , DecCBOR (Block h era)
+  , ToCBOR (Block h era)
+  , DecCBOR (Annotator (Block h era))
   ) =>
   Spec
 roundTripBlockSpec =
   prop (show (typeRep $ Proxy @(Block h era))) $ do
-    BaseQC.withNumTests 3 $ do
+    BaseQC.withNumTests 3 $
       conjoin
         [ roundTripEraExpectation @era @(Block h era)
         , roundTripAnnEraExpectation @era @(Block h era)